% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedForecast.R
\name{CombinedForecast}
\alias{CombinedForecast}
\title{Run cause-specific forecasts \lifecycle{experimental}}
\usage{
CombinedForecast(
  data,
  forecast_params,
  n_years = 2060 - 2016,
  folder_path = getwd()
)
}
\arguments{
\item{data}{Data table containing our processed cause-specific mortality rates}

\item{forecast_params}{Data table containing the control parameters for the cause specific forecasts}

\item{n_years}{Integer - the number of years to forecast into the future}

\item{folder_path}{where to save the outputs - defaults to the current working directory}
}
\value{
\item{mx_data_cause}{Data table containing the observed and forecast mortality rates by age, sex, IMD quintile and cause.}
\item{mx_data_all}{Data table containing the observed and forecast mortality rates for all causes combined.}
}
\description{
A wrapper that runs the forecast for each cause and groups the outputs together.
}
\examples{

\dontrun{

cforecast <- CombinedForecast(
  data = fread("tob_death_rates_national_2019-05-06_mort.tools_1.0.0.csv"),
  forecast_params = setDT(readxl::read_xlsx("disease specific mortality forecasting (for function input).xlsx"))
)

}


}
