% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp2013.R
\docType{data}
\name{esp2013}
\alias{esp2013}
\title{European Standard Population (ESP)}
\format{A data table}
\source{
https://www.causesofdeath.org/docs/standard.pdf
}
\usage{
esp2013
}
\description{
Weights for calculating age-standardized mortality rates for all ages
}
\keyword{datasets}
