% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupCauses.R
\name{GroupCauses}
\alias{GroupCauses}
\title{Aggregate ICD-10 codes by cause group}
\usage{
GroupCauses(
  data_file,
  lookup_file,
  strat_vars = c("year", "age", "sex", "imd_quintile", "laua")
)
}
\arguments{
\item{data_file}{Data table containing death counts ("n_deaths") by individual ICD10 codes.}

\item{lookup_file}{Data table showing how ICD-10 codes are grouped by cause.}

\item{strat_vars}{Character vector - the variables by which the dataset output should be stratified.}
}
\value{
Returns a data table of death counts by cause group and stratification variables.
}
\description{
Groups ICD-10 codes based on the groups of causes specified in our lists of diseases attributable
to tobacco, alcohol or both.
}
\details{
Identifies the ICD-10 codes that fall within each cause group.
For conditions that we consider only to occur in females (Breast, Cervix cancer, and hip fracture)
assign any occurance in males as 'other causes'. Finally, collapse the data by summing the death counts
within each cause group.

Before running this function you will need to run ExpandCodes() from the tobalcepi package.
}
\examples{

\dontrun{

data <- GroupCauses(
  data_file = copy(data),
  lookup_file = copy(lookups),
  strat_vars = c("year", "age", "sex", "imd_quintile")
)

}

}
