% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifetableSummary.R
\name{LifetableSummary}
\alias{LifetableSummary}
\title{Calculate lifetable summary statistics}
\usage{
LifetableSummary(mort_data, pop_data, strat_vars = c("age_cat", "sex",
  "imd_quintile", "year", "condition"), label = NULL)
}
\arguments{
\item{mort_data}{Data table - central rates of death in single years of age and period, stratified by cause.
Cause-specific death rates are "mix", stratified by subgroups of "age", "sex", "imd_quintile" and "year".}

\item{pop_data}{Data table - mid-year population sizes in single years of age and period. 
Counts must be in a variable named "N_pop".}

\item{strat_vars}{Character vector - the variables by which to stratify the summary.}

\item{label}{Character - the label to append to the calculated variables. Defaults to NULL.}
}
\value{
Returns a data table of the number of deaths and expected years of life lost by cause and subgroup.
}
\description{
Calculates key lifetable summary statistics from mortality rates and population numbers,
remaining life expectancy at each age and the expected number of years of life lost to death from each cause.
}
\details{
The function assumes that it is dealing with age and period in single years, so adjusting the size
of the time interval is not given as an option. The open age interval is 90+.
}
\examples{

\dontrun{
LifetableSummary(mort_data, pop_data)
}


}
