% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mort.tools-package.R
\docType{package}
\name{mort.tools-package}
\alias{mort.tools}
\alias{mort.tools-package}
\title{mort.tools: Process Mortality Microdata}
\description{
Read, Clean and Analyse Cause-Specific Mortality Data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stapm.gitlab.io/r-packages/mort.tools}
}

}
\author{
\strong{Maintainer}: Duncan Gillespie \email{duncan.gillespie@sheffield.ac.uk} (\href{https://orcid.org/0000-0003-3450-5747}{ORCID})

Authors:
\itemize{
  \item Laura Webster
  \item Colin Angus \email{c.r.angus@sheffield.ac.uk}
  \item Alan Brennan \email{a.brennan@sheffield.ac.uk}
}

}
\keyword{internal}
