% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalitySmooth.R
\name{MortalitySmooth}
\alias{MortalitySmooth}
\title{Smooth age and period surface of death rates \lifecycle{maturing}}
\usage{
MortalitySmooth(
  mort_data,
  condition_names,
  smooth_n_age = 3,
  smooth_n_year = 3
)
}
\arguments{
\item{mort_data}{Data table containing cause-specific mortality rates}

\item{condition_names}{names of the causes to include}

\item{smooth_n_age}{the age range for moving average - must be odd number}

\item{smooth_n_year}{the year range for moving average - must be odd number}
}
\value{
\item{mort_data}{Observed mortality}
\item{mort_data_smooth}{Smoothed observed mortality}
}
\description{
Applies a 2d moving average over age and period, by condition, sex and IMD quintile.
}
\examples{

\dontrun{

mx_smoothed <- MortalitySmooth(
  mort_data = test_data,
  condition_names = unique(test_data$condition),
  smooth_n_age = 5,
  smooth_n_year = 5
)

}


}
