% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSmooth.R
\name{MSmooth}
\alias{MSmooth}
\title{Smooth age and period surface of rates \lifecycle{maturing}}
\usage{
MSmooth(data, var_name, condition_names, smooth_n_age = 3, smooth_n_year = 3)
}
\arguments{
\item{data}{Data table containing cause-specific mortality rates}

\item{var_name}{Character - the name of the variable to be smoother}

\item{condition_names}{names of the causes to include}

\item{smooth_n_age}{the age range for moving average - must be odd number}

\item{smooth_n_year}{the year range for moving average - must be odd number}
}
\value{
\item{data}{Observed data}
\item{data_smooth}{Smoothed observed data}
}
\description{
Applies a 2d moving average over age and period, by condition, sex and IMD quintile.
}
\details{
Applied to mortality or morbidity data.
}
\examples{

\dontrun{

mx_smoothed <- MSmooth(
  data = test_data,
  var_name = "mix",
  condition_names = unique(test_data$condition),
  smooth_n_age = 5,
  smooth_n_year = 5
)

}


}
