% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{ReadData}
\alias{ReadData}
\title{Read ONS mortality microdata}
\usage{
ReadData(path)
}
\arguments{
\item{path}{Character - the path to the folder in which the data are stored.}
}
\value{
Returns a data table containing the data for all years.
}
\description{
Loads and cleans the microdata provided by the ONS into a clean form ready for further processing.
}
\details{
The ONS data was provided originally for years 2001-2014, with subsequent annual updates for 2015 and 2016.
The data is processed by an ONS analyst initially before being transferred to us.
}
\examples{

\dontrun{

data <- ReadData("D:/Death and population data")

}

}
