% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitOesoph.R
\name{SplitOesoph}
\alias{SplitOesoph}
\title{Split oesophageal (C15) ICD-10 codes in to SCC or AC \lifecycle{stable}}
\usage{
SplitOesoph(data, cruk_splits = oesoph_splits, n_deaths_var = "n_deaths")
}
\arguments{
\item{data}{Data table containing counts of death by cause.}

\item{cruk_splits}{Data table containing the percentages of oesphageal cancer that are SCC and AC.}

\item{n_deaths_var}{Character - the name of the variable containing the death counts.}
}
\value{
Returns a data table that is an updated version of data with extra rows added to reflect
the splitting of oesophageal cancer into two diseases.
}
\description{
Using the data provided to us by CRUK, split deaths marked as due to Oesophageal cancer into
deaths from the subtypes SCC and AC.
}
\examples{

\dontrun{

data <- SplitOesoph(data)

}

}
