% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tob_icd10_lookups.R
\docType{data}
\name{tob_icd10_lookups}
\alias{tob_icd10_lookups}
\title{Diseases related to tobacco and their ICD-10 defintions}
\format{A data table}
\source{
Our tobacco disease list in the folder PR_Disease_Risk_TA
}
\usage{
tob_icd10_lookups
}
\description{
Deaths in the mortality data are assigned a single ICD-10 code e.g. C00,
but for our analysis we need to group these into cause-groups according to groups e.g. C00-C06.
To do this, we define each group by its consituent ICD-10 codes e.g. C00, C01, C02, C03, C04, C05, C06.
}
\keyword{datasets}
