% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadPopData.R
\name{ReadPopData}
\alias{ReadPopData}
\title{Load the population counts data}
\usage{
ReadPopData(path)
}
\arguments{
\item{path}{Character - the path to the folder in which the data are stored.}
}
\value{
Returns a data table containing the data for all years.
}
\description{
Reads and cleans the population count data by age, year, sex and IMD quintile.
}
\examples{

\dontrun{

data <- ReadPopData("D:/Death and population data")

}

}
